+++ src/lib/breadcrumb-tooltip/index.ts
@@ -1,0 +1,3 @@
+export * from './breadcrumb-tooltip.types';
+export * from './breadcrumb-tooltip.token';
+export * from './breadcrumb-tooltip.provider';