+++ src/public-api.ts
@@ -2,6 +2,8 @@
  * Public API Surface of breadcrumbs
  */
 export * from './lib/breadcrumbs.module';
+export * from './lib/breadcrumb-tooltip';
 export * from './lib/components/breadcrumb/breadcrumb.component';
 export * from './lib/directives/breadcrumb-item.directive';
+export * from './lib/directives/breadcrumb-label.directive';
 export * from './lib/services/breadcrumbs.service';